Move remote models to remote package
