+++ src/main/java/io/jenkins/plugins/appcenter/model/remote/BuildInfo.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.model;
+package io.jenkins.plugins.appcenter.model.remote;
 
 public class BuildInfo {
     public final String branch;
