+++ src/main/java/io/jenkins/plugins/appcenter/model/remote/ReleaseDetailsUpdateResponse.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.model;
+package io.jenkins.plugins.appcenter.model.remote;
 
 public class ReleaseDetailsUpdateResponse {
     public final String release_notes;
