+++ src/main/java/io/jenkins/plugins/appcenter/model/remote/ReleaseUploadBeginResponse.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.model;
+package io.jenkins.plugins.appcenter.model.remote;
 
 public final class ReleaseUploadBeginResponse {
     public final String upload_id;
