+++ src/main/java/io/jenkins/plugins/appcenter/model/remote/Status.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.model;
+package io.jenkins.plugins.appcenter.model.remote;
 
 public enum Status {
     committed,
