+++ src/main/java/io/jenkins/plugins/appcenter/task/UploadTask.java
@@ -4,13 +4,13 @@ import hudson.FilePath;
 import hudson.model.TaskListener;
 import io.jenkins.plugins.appcenter.AppCenterException;
 import io.jenkins.plugins.appcenter.model.AppCenterServiceFactory;
-import io.jenkins.plugins.appcenter.model.DestinationId;
-import io.jenkins.plugins.appcenter.model.ReleaseDetailsUpdateRequest;
-import io.jenkins.plugins.appcenter.model.ReleaseDetailsUpdateResponse;
-import io.jenkins.plugins.appcenter.model.ReleaseUploadBeginResponse;
-import io.jenkins.plugins.appcenter.model.ReleaseUploadEndRequest;
-import io.jenkins.plugins.appcenter.model.ReleaseUploadEndResponse;
-import io.jenkins.plugins.appcenter.model.Status;
+import io.jenkins.plugins.appcenter.model.remote.DestinationId;
+import io.jenkins.plugins.appcenter.model.remote.ReleaseDetailsUpdateRequest;
+import io.jenkins.plugins.appcenter.model.remote.ReleaseDetailsUpdateResponse;
+import io.jenkins.plugins.appcenter.model.remote.ReleaseUploadBeginResponse;
+import io.jenkins.plugins.appcenter.model.remote.ReleaseUploadEndRequest;
+import io.jenkins.plugins.appcenter.model.remote.ReleaseUploadEndResponse;
+import io.jenkins.plugins.appcenter.model.remote.Status;
 import okhttp3.MediaType;
 import okhttp3.MultipartBody;
 import okhttp3.RequestBody;
